COMP: Eliminate or suppress compiler warnings on the 64-bit Windows builds.

+++ Utilities/vtkexodus2/expblk.c
@@ -436,7 +436,8 @@ int ex_put_block( int         exoid,
 
 /* store element type as attribute of connectivity variable */
 
-   if ((ncattput (exoid, connid, ATT_NAME_ELB, NC_CHAR, strlen(entry_descrip)+1, 
+   if ((ncattput (exoid, connid, ATT_NAME_ELB, NC_CHAR,
+             (int)strlen(entry_descrip)+1,
              (void*) entry_descrip)) == -1)
    {
      exerrval = ncerr;
